Fix endian bug, also add sanity checks.
There was a bug that caused an endian-read problem. I also added a sanity check placing a hard limit on how big the queue file could grow. Because of the variables involved, it would be limited to 4GB anyway, but on a small device that would be pretty bad. I've put a 2GB limit in for now, but we can discuss whether it has merit.