Demo how tape can be used to retry tasks
Update tape-sample with some basic retry functionality to demonstrate a
common use case for the library.

I have updated the onFailure callback method a little to retry the last task.
To simulate some random failures I've fudged the API key randomly before making
the post to imgur.

The number of retries are then shown back in the main activity alongside the
image URL.