Replace RuntimeException with new UncheckedIOException
This is the same as Java 8's [UncheckedIOException](https://docs.oracle.com/javase/8/docs/api/java/io/UncheckedIOException.html). We've copied it so that we can maintain compatibility with Java 7.

Other options:
* Continue to throw RuntimeException with a nicer error message (i.e. remove the "todo")
* Create a custom exception Tape class like we had in tape v1 (ref #154)

I think this is nicer than the second as it keeps the ergonomics of Java 8 for those stuck on having to support Java 7.