Closes #198: Implement the BlockingQueue as a wrapper around the QueueFile
QueueFile is now an ObjectQueue<byte[]> and can thus be used without overhead in the blocking queue. ObjectQueue's file():QueueFile method has been moved to FileObjectQueue where it fits better as InMemoryObjectQueue doesn't have a backing file.
Removed the ByteArrayConverter again as it was the initial attempt to optimize for byte[] queues.